+++ model/deep_survival_image_only_attention/deep_survival_image_only_attention.py
